Convert all of Molecules to Typescript and create custom typings folder
I went through the Molecules Components and converted each to Typescript.

In the Music and Flag compoent I removed the event parameter from the play pause function
I also had some modifiers on the Icon button that were not able to be used and were not in effect

The react-simply-carousel component had no typescript declorations.
I was unable to install the types using NPM so I created a typing folder and declared the
package in there.